move tests for manual scaler
