+++ src/schematic/component_test.rs
@@ -605,7 +605,7 @@ fn test_to_pod_spec_with_policy() {
                 .get("kubernetes.io/arch")
                 .expect("an arch should be present")
         );
-        
+
         pod.containers.clone().iter().for_each(|c| {
             assert_eq!(Some("Always".to_string()), c.image_pull_policy);
         });
@@ -713,7 +713,10 @@ fn test_to_deployment_spec() {
     );
     let comp = comp_res.as_ref().expect("component should exist");
     let mut labels = BTreeMap::new();
-    labels.insert("app.kubernetes.io/name".to_string(), "test_deploy".to_string());
+    labels.insert(
+        "app.kubernetes.io/name".to_string(),
+        "test_deploy".to_string(),
+    );
     let resloved_val =
         resolve_parameters(comp.parameters.clone(), BTreeMap::new()).expect("resolved parameter");
     let deploy = comp.to_deployment_spec(resloved_val, Some(labels.clone()), None);
